Fixing some more prototypes. Pointed out by Markku Laukkanen.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@2597 06827809-a52a-0410-b366-d66718629ded

+++ net-snmp/agent/mibgroup/ucd-snmp/vmstat.c
@@ -221,9 +221,9 @@ unsigned vmstat (int iindex)
 
 unsigned char *var_extensible_vmstat(struct variable *vp,
 				     oid *name,
-				     int *length,
+				     size_t *length,
 				     int exact,
-				     int *var_len,
+				     size_t *var_len,
 				     WriteMethod **write_method)
 {
